+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.7.0] = 2025-02-25
+
+### Added
+
+- Support sending query params to `Customers.CreateAsync` and `Customers.UpdateAsync` methods.
+- Added fields to `Customer` model: `SatValidatedAt`, `EditLink` and `EditLinkExpiresAt`.
+- Added targets for .NET 6.0 and .NET 7.0.
+
 ## [4.6.0] - 2024-23-09
 
 ### Added